Refactoring createReview page and separate logic in CreateREviewCompoennt
